@@ -4,18 +4,22 @@
   "private": true,
   "scripts": {
     "lint": "eslint --ext .ts,.tsx .",
-    "dev": "craco start",
-    "build": "craco build",
-    "test": "craco test"
+    "dev": "NODE_OPTIONS='--openssl-legacy-provider' craco start",
+    "build": "NODE_OPTIONS='--openssl-legacy-provider' craco build",
+    "test": "NODE_OPTIONS='--openssl-legacy-provider' craco test"
   },
   "dependencies": {
     "@airgap/beacon-sdk": "^3.3.0",
-    "@craco/craco": "^6.2.0",
-    "@fortawesome/fontawesome-svg-core": "^1.2.36",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.15",
-    "@onflow/fcl": "^1.0.2",
+    "@commitlint/cli": "^13.1.0",
+    "@commitlint/config-conventional": "^13.1.0",
+    "@craco/craco": "^7.1.0",
+    "@fortawesome/fontawesome-svg-core": "^6.4.0",
+    "@fortawesome/free-brands-svg-icons": "^6.4.0",
+    "@fortawesome/free-solid-svg-icons": "^6.4.0",
+    "@fortawesome/react-fontawesome": "^0.2.0",
+    "@mysten/sui.js": "^0.37.1",
+    "@onflow/fcl": "^1.4.1",
+    "@suiet/wallet-kit": "^0.2.15",
     "@taquito/beacon-wallet": "^15.0.0",
     "@taquito/taquito": "^15.0.0",
     "@taquito/tzip12": "^15.0.0",
@@ -24,37 +28,44 @@
     "@testing-library/user-event": "^12.1.10",
     "@web3-react/core": "^6.1.9",
     "@web3-react/injected-connector": "^6.0.7",
+    "autoprefixer": "^9.8.6",
     "bignumber.js": "^9.1.1",
     "deep-equal": "^2.0.5",
     "ethers": "5.4.1",
     "firebase": "^9.1.2",
+    "firebase-tools": "^9.16.5",
+    "husky": "^7.0.1",
     "keccak256": "^1.0.6",
     "merkletreejs": "^0.2.27",
     "p5": "^1.4.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "prettier": "^2.3.2",
+    "prettier-plugin-solidity": "^1.0.0-beta.17",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-gif-player": "^0.4.2",
     "react-helmet": "^6.1.0",
     "react-hooks-use-modal": "^2.1.0",
     "react-p5": "^1.3.19",
     "react-p5-wrapper": "^2.4.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^5.0.1",
+    "stream-browserify": "^3.0.0",
     "svg64": "^1.2.0",
-    "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
   "devDependencies": {
     "@types/deep-equal": "^1.0.1",
     "@types/jest": "^26.0.15",
-    "@types/node": "^12.0.0",
-    "@types/react": "^17.0.0",
-    "@types/react-dom": "^17.0.0",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router-dom": "^5.1.8",
-    "autoprefixer": "^9",
-    "postcss": "^7",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat"
+    "@types/node": "^20.4.2",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
+    "@types/react-helmet": "^6.1.6",
+    "@types/react-router-dom": "^5.3.3",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.26",
+    "postcss-cli": "^10.1.0",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.1.6"
   },
   "eslintConfig": {
     "extends": [
