@@ -1,7 +1,18 @@
 module.exports = {
   style: {
-    postcss: {
-      plugins: [require("tailwindcss"), require("autoprefixer")],
+    // why use postcssOptions? -> https://github.com/dilanx/craco/issues/353
+    postcssOptions: {
+        plugins: [
+            require('tailwindcss'),
+            require('autoprefixer'),
+        ],
+    },
+  },
+  webpack: {
+    configure: {
+      resolve: {
+        fallback: { "stream": require.resolve("stream-browserify") }
+      }
     },
   },
 };
