@@ -1,23 +1,13 @@
+/** @type {import('tailwindcss').Config} */
 module.exports = {
-  purge: ['./src/pages/**/*.{js,ts,jsx,tsx}',"./src/components/**/*.tsx"],
+  // content: ["./src/**/*.{js,jsx,ts,tsx}", "./public/index.html"],
+  purge: ['./src/pages/**/*.{js,ts,jsx,tsx}',"./src/components/**/*.tsx","./public/index.html"],
   darkMode: false, // or 'media' or 'class'
   theme: {
-    extend: {
-      colors: {
-        "marimo-1": "#1C203C",
-        "marimo-2": "#17207F",
-        "marimo-3": "#234F90",
-        "marimo-4": "#408D9D",
-        "marimo-5": "#60A5FA",
-        "marimo-6": "#4B5563",
-      },
-    },
+    extend: {},
     fontSize: {
       "2xs": "0.5em",
     },
-  },
-  variants: {
-    extend: {},
   },
   plugins: [],
 };
